+++ backend/src/main/java/edonymyeon/backend/auth/annotation/AuthPrincipal.java
@@ -1,0 +1,12 @@
+package edonymyeon.backend.auth.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AuthPrincipal {
+
+}
