+++ backend/src/main/java/edonymyeon/backend/auth/application/dto/LoginRequest.java
@@ -1,0 +1,5 @@
+package edonymyeon.backend.auth.application.dto;
+
+public record LoginRequest(String email, String password) {
+
+}
