--- backend/src/main/java/edonymyeon/backend/config/WebMvcConfiguration.java
@@ -1,20 +1,0 @@
-package edonymyeon.backend.config;
-
-import edonymyeon.backend.ui.argumentresolver.MemberArgumentResolver;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@RequiredArgsConstructor
-@Configuration
-public class WebMvcConfiguration implements WebMvcConfigurer {
-
-    private final MemberArgumentResolver memberArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(memberArgumentResolver);
-    }
-}
