--- backend/src/main/java/edonymyeon/backend/controlleradvice/dto/ExceptionResponse.java
@@ -1,4 +1,0 @@
-package edonymyeon.backend.controlleradvice.dto;
-
-public record ExceptionResponse(int errorCode, String errorMessage) {
-}
