--- backend/src/main/java/edonymyeon/backend/domain/ProfileImageInfo.java
@@ -1,13 +1,0 @@
-package edonymyeon.backend.domain;
-
-import jakarta.persistence.Entity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class ProfileImageInfo extends ImageInfo {
-
-}
