+++ backend/src/main/java/edonymyeon/backend/global/config/WebMvcConfiguration.java
@@ -1,0 +1,20 @@
+package edonymyeon.backend.global.config;
+
+import edonymyeon.backend.auth.ui.argumentresolver.AuthArgumentResolver;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@RequiredArgsConstructor
+@Configuration
+public class WebMvcConfiguration implements WebMvcConfigurer {
+
+    private final AuthArgumentResolver authArgumentResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(authArgumentResolver);
+    }
+}
