+++ backend/src/main/java/edonymyeon/backend/global/controlleradvice/dto/ExceptionResponse.java
@@ -1,0 +1,5 @@
+package edonymyeon.backend.global.controlleradvice.dto;
+
+public record ExceptionResponse(int errorCode, String errorMessage) {
+
+}
