+++ backend/src/main/java/edonymyeon/backend/global/exception/EdonymyeonException.java
@@ -1,0 +1,23 @@
+package edonymyeon.backend.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class EdonymyeonException extends RuntimeException {
+
+    private final ExceptionInformation exceptionInformation;
+
+    public EdonymyeonException(ExceptionInformation exceptionInformation) {
+        super();
+        this.exceptionInformation = exceptionInformation;
+    }
+
+    public int getCode() {
+        return exceptionInformation.getCode();
+    }
+
+    @Override
+    public String getMessage() {
+        return exceptionInformation.getMessage();
+    }
+}
