+++ backend/src/main/java/edonymyeon/backend/image/ImageFileNameStrategy.java
@@ -1,0 +1,6 @@
+package edonymyeon.backend.image;
+
+public interface ImageFileNameStrategy {
+
+    String createName(String originalFileName);
+}
