+++ backend/src/main/java/edonymyeon/backend/image/postimage/PostImageInfoRepository.java
@@ -1,0 +1,10 @@
+package edonymyeon.backend.image.postimage;
+
+import edonymyeon.backend.image.postimage.domain.PostImageInfo;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostImageInfoRepository extends JpaRepository<PostImageInfo, Long> {
+
+    List<PostImageInfo> findAllByPostId(Long postId);
+}
