+++ backend/src/main/java/edonymyeon/backend/image/postimage/ProfileImageInfo.java
@@ -1,0 +1,14 @@
+package edonymyeon.backend.image.postimage;
+
+import edonymyeon.backend.image.domain.ImageInfo;
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class ProfileImageInfo extends ImageInfo {
+
+}
