+++ backend/src/main/java/edonymyeon/backend/member/application/dto/MemberIdDto.java
@@ -1,0 +1,5 @@
+package edonymyeon.backend.member.application.dto;
+
+public record MemberIdDto(Long id) {
+
+}
