+++ backend/src/main/java/edonymyeon/backend/member/domain/Member.java
@@ -1,0 +1,40 @@
+package edonymyeon.backend.member.domain;
+
+import edonymyeon.backend.image.postimage.ProfileImageInfo;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String email;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private String nickname;
+
+    private String introduction;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn
+    private ProfileImageInfo profileImageInfo;
+    //todo: 검증 로직
+}
