+++ backend/src/main/java/edonymyeon/backend/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package edonymyeon.backend.member.repository;
+
+import edonymyeon.backend.member.domain.Member;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByEmailAndPassword(final String email, final String password);
+}
