+++ backend/src/main/java/edonymyeon/backend/post/repository/PostRepository.java
@@ -1,0 +1,8 @@
+package edonymyeon.backend.post.repository;
+
+import edonymyeon.backend.post.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+
+}
