+++ backend/src/main/java/edonymyeon/backend/post/ui/PostController.java
@@ -1,0 +1,30 @@
+package edonymyeon.backend.post.ui;
+
+import edonymyeon.backend.auth.annotation.AuthPrincipal;
+import edonymyeon.backend.member.application.dto.MemberIdDto;
+import edonymyeon.backend.post.application.PostService;
+import edonymyeon.backend.post.application.dto.PostRequest;
+import edonymyeon.backend.post.application.dto.PostResponse;
+import java.net.URI;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RequestMapping("/posts")
+@RestController
+public class PostController {
+
+    private final PostService postService;
+
+    @PostMapping
+    public ResponseEntity<PostResponse> createPost(@AuthPrincipal MemberIdDto memberId,
+                                                   @ModelAttribute PostRequest postRequest) {
+        PostResponse response = postService.createPost(postRequest);
+        return ResponseEntity.created(URI.create("/posts/" + response.id()))
+                .body(response);
+    }
+}
