--- backend/src/main/java/edonymyeon/backend/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package edonymyeon.backend.repository;
-
-import edonymyeon.backend.domain.Member;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmailAndPassword(final String email, final String password);
-}
