--- backend/src/main/java/edonymyeon/backend/repository/PostImageInfoRepository.java
@@ -1,10 +1,0 @@
-package edonymyeon.backend.repository;
-
-import edonymyeon.backend.domain.PostImageInfo;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostImageInfoRepository extends JpaRepository<PostImageInfo, Long> {
-
-    List<PostImageInfo> findAllByPostId(Long postId);
-}
