--- backend/src/main/java/edonymyeon/backend/repository/PostRepository.java
@@ -1,7 +1,0 @@
-package edonymyeon.backend.repository;
-
-import edonymyeon.backend.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-}
