--- backend/src/main/java/edonymyeon/backend/service/ImageFileNameStrategy.java
@@ -1,6 +1,0 @@
-package edonymyeon.backend.service;
-
-public interface ImageFileNameStrategy {
-
-    String createName(String originalFileName);
-}
