--- backend/src/main/java/edonymyeon/backend/service/request/LoginRequest.java
@@ -1,5 +1,0 @@
-package edonymyeon.backend.service.request;
-
-public record LoginRequest(String email, String password) {
-
-}
