--- backend/src/main/java/edonymyeon/backend/service/request/PostRequest.java
@@ -1,14 +1,0 @@
-package edonymyeon.backend.service.request;
-
-import java.util.List;
-import org.springframework.web.multipart.MultipartFile;
-
-public record PostRequest(
-        String title,
-        String content,
-        Long price,
-        List<MultipartFile> images
-) {
-
-}
-
