--- backend/src/main/java/edonymyeon/backend/service/response/MemberIdDto.java
@@ -1,5 +1,0 @@
-package edonymyeon.backend.service.response;
-
-public record MemberIdDto(Long id) {
-
-}
