--- backend/src/main/java/edonymyeon/backend/service/response/PostResponse.java
@@ -1,4 +1,0 @@
-package edonymyeon.backend.service.response;
-
-public record PostResponse(Long id) {
-}
