--- backend/src/main/java/edonymyeon/backend/ui/annotation/AuthenticationPrincipal.java
@@ -1,12 +1,0 @@
-package edonymyeon.backend.ui.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AuthenticationPrincipal {
-
-}
