+++ backend/src/test/java/edonymyeon/backend/TestConfig.java
@@ -1,7 +1,7 @@
 package edonymyeon.backend;
 
+import edonymyeon.backend.image.ImageFileUploader;
 import edonymyeon.backend.service.ImageFileNameMockStrategy;
-import edonymyeon.backend.service.ImageFileUploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -9,6 +9,7 @@ import org.springframework.context.annotation.Bean;
 @RequiredArgsConstructor
 @TestConfiguration
 public class TestConfig {
+
     @Bean
     public ImageFileUploader imageFileUploader() {
         return new ImageFileUploader(new ImageFileNameMockStrategy(
